+++ examples/scripts/getting_started/mcmc_example.py
@@ -1,7 +1,6 @@
 import sys
 
 import numpy as np
-import pybamm
 
 import pybop
 
@@ -9,7 +8,6 @@ import pybop
 parallel = True if sys.platform != "win32" else False
 
 # Parameter set and model definition
-solver = pybamm.IDAKLUSolver()
 parameter_set = pybop.ParameterSet("Chen2020")
 parameter_set.update(
     {
@@ -17,7 +15,7 @@ parameter_set.update(
         "Positive electrode active material volume fraction": 0.71,
     }
 )
-synth_model = pybop.lithium_ion.DFN(parameter_set=parameter_set, solver=solver)
+synth_model = pybop.lithium_ion.SPMe(parameter_set=parameter_set)
 
 # Fitting parameters
 parameters = pybop.Parameters(
@@ -38,7 +36,7 @@ init_soc = 0.5
 sigma = 0.005
 experiment = pybop.Experiment(
     [
-        ("Discharge at 0.5C for 6 minutes (5 second period)",),
+        ("Discharge at 0.5C for 3 minutes (5 second period)",),
     ]
 )
 values = synth_model.predict(
@@ -56,14 +54,11 @@ dataset = pybop.Dataset(
         "Time [s]": values["Time [s]"].data,
         "Current function [A]": values["Current [A]"].data,
         "Voltage [V]": noisy(values["Voltage [V]"].data, sigma),
-        "Bulk open-circuit voltage [V]": noisy(
-            values["Bulk open-circuit voltage [V]"].data, sigma
-        ),
     }
 )
 
-model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=pybamm.IDAKLUSolver())
-signal = ["Voltage [V]", "Bulk open-circuit voltage [V]"]
+model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
+signal = ["Voltage [V]"]
 
 # Generate problem, likelihood, and sampler
 problem = pybop.FittingProblem(
@@ -75,7 +70,7 @@ posterior = pybop.LogPosterior(likelihood)
 sampler = pybop.DifferentialEvolutionMCMC(
     posterior,
     chains=3,
-    max_iterations=300,
+    max_iterations=250,  # Reduced for CI, increase for improved posteriors
     warm_up=100,
     verbose=True,
     parallel=parallel,  # (macOS/WSL/Linux only)