+++ tests/integration/test_model_experiment_changes.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pybamm
 import pytest
 
 import pybop
@@ -36,13 +37,18 @@ class TestModelAndExperimentChanges:
     def parameters(self, request):
         return request.param
 
-    def test_changing_experiment(self, parameters):
+    @pytest.fixture
+    def solver(self):
+        return pybamm.IDAKLUSolver(atol=1e-6, rtol=1e-6)
+
+    @pytest.mark.integration
+    def test_changing_experiment(self, parameters, solver):
         # Change the experiment and check that the results are different.
 
         parameter_set = pybop.ParameterSet("Chen2020")
         parameter_set.update(parameters.as_dict("true"))
         initial_state = {"Initial SoC": 0.5}
-        model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
+        model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
 
         t_eval = np.arange(0, 3600, 2)  # Default 1C discharge to cut-off voltage
         solution_1 = model.predict(initial_state=initial_state, t_eval=t_eval)
@@ -66,19 +72,19 @@ class TestModelAndExperimentChanges:
         np.testing.assert_allclose(cost_1, 0, atol=1e-5)
         np.testing.assert_allclose(cost_2, 0, atol=1e-5)
 
-    def test_changing_model(self, parameters):
+    def test_changing_model(self, parameters, solver):
         # Change the model and check that the results are different.
 
         parameter_set = pybop.ParameterSet("Chen2020")
         parameter_set.update(parameters.as_dict("true"))
         initial_state = {"Initial SoC": 0.5}
-        experiment = pybop.Experiment(["Charge at 1C until 4.1 V (2 seconds period)"])
+        experiment = pybop.Experiment(["Charge at 1C until 4.1 V (30 seconds period)"])
 
-        model = pybop.lithium_ion.SPM(parameter_set=parameter_set)
+        model = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
         solution_1 = model.predict(initial_state=initial_state, experiment=experiment)
         cost_1 = self.final_cost(solution_1, model, parameters)
 
-        model = pybop.lithium_ion.SPMe(parameter_set=parameter_set)
+        model = pybop.lithium_ion.SPMe(parameter_set=parameter_set, solver=solver)
         solution_2 = model.predict(initial_state=initial_state, experiment=experiment)
         cost_2 = self.final_cost(solution_2, model, parameters)
 
@@ -104,11 +110,11 @@ class TestModelAndExperimentChanges:
         signal = ["Voltage [V]"]
         problem = pybop.FittingProblem(model, parameters, dataset, signal=signal)
         cost = pybop.RootMeanSquaredError(problem)
-        optim = pybop.PSO(cost)
+        optim = pybop.NelderMead(cost)
         results = optim.run()
         return results.final_cost
 
-    def test_multi_fitting_problem(self):
+    def test_multi_fitting_problem(self, solver):
         parameter_set = pybop.ParameterSet("Chen2020")
         parameters = pybop.Parameters(
             pybop.Parameter(
@@ -120,7 +126,7 @@ class TestModelAndExperimentChanges:
             )
         )
 
-        model_1 = pybop.lithium_ion.SPM(parameter_set=parameter_set)
+        model_1 = pybop.lithium_ion.SPM(parameter_set=parameter_set, solver=solver)
         experiment_1 = pybop.Experiment(
             ["Discharge at 0.5C for 5 minutes (10 seconds period)"]
         )
@@ -133,7 +139,9 @@ class TestModelAndExperimentChanges:
             }
         )
 
-        model_2 = pybop.lithium_ion.SPMe(parameter_set=parameter_set.copy())
+        model_2 = pybop.lithium_ion.SPMe(
+            parameter_set=parameter_set.copy(), solver=solver
+        )
         experiment_2 = pybop.Experiment(
             ["Discharge at 1C for 3 minutes (10 seconds period)"]
         )